// frontend/generated/models/PromotionalCampaignDto.ts
export interface PromotionalCampaignDto {
    /**
     * 
     * @type {number}
     * @memberof PromotionalCampaignDto
     */
    campaignId?: number;
    /**
     * 
     * @type {string}
     * @memberof PromotionalCampaignDto
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof PromotionalCampaignDto
     */
    description?: string;
    /**
     * 
     * @type {Date}
     * @memberof PromotionalCampaignDto
     */
    startDate?: Date;
    /**
     * 
     * @type {Date}
     * @memberof PromotionalCampaignDto
     */
    endDate?: Date;
    /**
     * 
     * @type {number}
     * @memberof PromotionalCampaignDto
     */
    discountRate?: number;
    /**
     * 
     * @type {boolean}
     * @memberof PromotionalCampaignDto
     */
    isVisible?: boolean;
    /**
     * 
     * @type {Array<BlindBoxCampaignDto>}
     * @memberof PromotionalCampaignDto
     */
    blindBoxCampaigns?: Array<BlindBoxCampaignDto>;
    /**
     * 
     * @type {Date}
     * @memberof PromotionalCampaignDto
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof PromotionalCampaignDto
     */
    updatedAt?: Date;
}